+++ js/image-effects.js
@@ -1,3 +1,6 @@
+// image-effects.js
+import { updateScale, onSmallerClick, onBiggerClick } from './image-utils.js';
+
 // Объект, содержащий эффекты и их соответствующие CSS-фильтры
 const effects = {
   none: { name: 'none', filter: 'none' },
@@ -41,48 +44,9 @@ const updateSliderValue = (value) => {
   }
 };
 
-// Масштабирование изображения
-const SCALE_STEP = 0.25;
-const MIN_SCALE = 0.25;
-const MAX_SCALE = 1;
-
-let currentScale = MAX_SCALE;
-
-const hiddenScaleInput = document.createElement('input');
-hiddenScaleInput.type = 'hidden';
-hiddenScaleInput.name = 'scale';
-document.querySelector('.img-upload__form').appendChild(hiddenScaleInput);
-
-const updateScale = () => {
-  const imgElement = document.querySelector('.img-upload__preview img');
-
-  if (imgElement) {
-    imgElement.style.transform = `scale(${currentScale})`;
-  }
-
-  const scaleControlValue = document.querySelector('.scale__control--value');
-
-  scaleControlValue.value = `${currentScale * 100}%`;
-  hiddenScaleInput.value = currentScale;
-};
-
-const onSmallerClick = () => {
-  if (currentScale > MIN_SCALE) {
-    currentScale -= SCALE_STEP;
-    updateScale();
-  }
-};
-
-const onBiggerClick = () => {
-  if (currentScale < MAX_SCALE) {
-    currentScale += SCALE_STEP;
-    updateScale();
-  }
-};
-
 // Добавление слушателей событий
 document.querySelector('.effects__list').addEventListener('change', onEffectChange);
 document.querySelector('.scale__control--smaller').addEventListener('click', onSmallerClick);
 document.querySelector('.scale__control--bigger').addEventListener('click', onBiggerClick);
 
-export { onEffectChange, updateScale, onSmallerClick, onBiggerClick };
+export { onEffectChange };
